Added version class to verify build.xml
This is the initial build and uses ant script to build test and compile
one class which provides basic Version information.

This shouldn't stay this way as the version setting needs to be managed
in script (probably ant), however it does prove the basic ant script is
a good starting point.

I know nothing about Maven and wanted to avoid any IDE build scripts but
imagine I will regret this adecision later.