Configuration is a singleton to act as a global
There should be just one configuration acting at one time so by creating
it as a Singleton it means it is only ever created once.

It is retrieved through the Configuration.getInstance() method.

For testing purposes we have included a Configuration.delete() method
which deletes the configuration and allows a new one to be retrieved.