able to retrieve the default Hostname
This was really created to ensure a call could be made to the p6
database.  It has highlighted the need to be clear about what I am
testing and as such the test is not right.

It isn't right because it calls the Primavera version of the class
whereas it should just be calling the RMIURL class belonging to this
library.

In developing this library I need to ensure I am testing the library
calling the right P6 API and not calling direct myself.  I don't really
want to use a real database to keep speed down but I also have to ensure
I am genuinely testing.

This is going to be hard.