capable of testing for local, RMI or WS
The environment needs to be setup differently for Local, RMI and Web
Services so I want to be able to exclude RMI and WS specific classes
when I am testing the Local connection, Local and WS when testing RMI
and RMI and Local when testing WS.

This is the first stage to do that.

I am going to build a few classes for Local and when completed, I will
repeat the same for RMI and finally Web Services.  This will require the
ability to make use of specific config file for testing as well as
setting up the environment using ant, which will be a learning
experience to say the least.

Having resolved the build scripts I can concentrate on implementing the
Java API for Web Services, since the RMI should not change.

This will be the difficult value add part.