initial insufficient test for RMIURL
This is insufficient because it can only work with Local or RMI
connections and not Web Services.  It was really only done to help build
the ant script on the one hand and learn what RMIURL really does.

To work properly my RMIURL class has to wrap around the appropriate
concrete class for the underlying connectivity (Java API or Web
Services)