ability to choose Java API or Web Service
This is the first time I have done this and I have introduced some
conventions.

The concrete class called by the client application has the same name as
the Java API class it is re-implementing and uses an interface with the
same name prefixed by "I" to reference the Java API or Web Service
class.  The Java API class is the same name prefixed by "J" and the web
service is the same name prefixed by "W".

Therefore:

RMIURL is the name of the class called by the client application and
matches the name in the Java API.

IRMIURL is the name of the interface implemented by the concrete SDK
classes.

JRMIURL is the name of the Java API and Java RMI class.

WRMIURL is the name of the Web Services class.

Worth noting in this case the WRMIURL class does nothing practical and
is not exercised in any testing to date.  The JRMIURL class just wraps
around the corresponding Java API class and this will be the case for
Java API classes every time.  Easy and boring to implement.